Figured out method to get album image
