+++ app/api/start/[id]/route.ts
@@ -4,7 +4,7 @@ import { NextResponse } from 'next/server';
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
   try {
-    const { id } = await params
+    const { id } = params
     const genre = genres.data.find(item => item.id === id)
 
     if (!genre) {
