+++ lib/youtube.tsx
@@ -1,0 +1,14 @@
+import Youtube, { YouTubeProps } from 'react-youtube';
+
+export function getYoutube({ url }: { url: any }) {
+  const onPlayerReady: YouTubeProps['onReady'] = (event) => {
+    event.target.pauseVideo();
+  }
+
+  const opts: YouTubeProps['opts'] = {
+    height: '360',
+    width: '480'
+  }
+
+  return <Youtube videoId={url} opts={opts} onReady={onPlayerReady} />;
+}
