+++ next.config.ts
@@ -1,7 +1,17 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
-};
+  images: {
+      remotePatterns: [
+        {
+          protocol: 'http',
+          hostname: 'img.youtube.com',
+          port: '',
+          pathname: '/vi/**',
+          search: '',
+        },
+      ],
+    },
+}
 
 export default nextConfig;
